+++ tests/DateTimeProvider/MutableDateTimeProviderTest.phpt
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Test: Kdyby\DateTimeProvider\MutableProvider.
+ * Test: Kdyby\DateTimeProvider\MutableDateTimeProvider.
  *
  * @testCase
  */
@@ -11,17 +11,17 @@ declare(strict_types = 1);
 namespace KdybyTests\DateTimeProvider;
 
 use DateTimeImmutable;
-use Kdyby\DateTimeProvider\Provider\MutableProvider;
+use Kdyby\DateTimeProvider\Provider\MutableDateTimeProvider;
 use Tester\Assert;
 
 require_once __DIR__ . '/../bootstrap.php';
 
-class MutableProviderTest extends \Tester\TestCase
+class MutableDateTimeProviderTest extends \Tester\TestCase
 {
 
 	public function testConstant(): void
 	{
-		$tp = new MutableProvider(new DateTimeImmutable('2013-09-14 03:53:21'));
+		$tp = new MutableDateTimeProvider(new DateTimeImmutable('2013-09-14 03:53:21'));
 		$datetime = $tp->getDateTime();
 		$date = $tp->getDate();
 		$time = $tp->getTime();
@@ -39,20 +39,20 @@ class MutableProviderTest extends \Tester\TestCase
 	{
 		date_default_timezone_set('Europe/Prague');
 
-		$tp = new MutableProvider(new DateTimeImmutable(date('Y-m-d H:i:s', 1379123601)));
+		$tp = new MutableDateTimeProvider(new DateTimeImmutable(date('Y-m-d H:i:s', 1379123601)));
 		Assert::same('Europe/Prague', $tp->getTimeZone()->getName());
 		Assert::same('2013-09-14 03:53:21 +02:00', $tp->getDateTime()->format('Y-m-d H:i:s P'));
 
 		date_default_timezone_set('Europe/London');
 
-		$tp = new MutableProvider(new DateTimeImmutable(date('Y-m-d H:i:s', 1379123601)));
+		$tp = new MutableDateTimeProvider(new DateTimeImmutable(date('Y-m-d H:i:s', 1379123601)));
 		Assert::same('Europe/London', $tp->getTimeZone()->getName());
 		Assert::same('2013-09-14 02:53:21 +01:00', $tp->getDateTime()->format('Y-m-d H:i:s P'));
 	}
 
 	public function testChangePrototype(): void
 	{
-		$tp = new MutableProvider($originalTime = new DateTimeImmutable('2013-09-14 03:53:21'));
+		$tp = new MutableDateTimeProvider($originalTime = new DateTimeImmutable('2013-09-14 03:53:21'));
 		Assert::same($originalTime, $tp->getDateTime());
 
 		$tp->changePrototype($changedTime = new DateTimeImmutable('2015-01-09 18:34:00'));
@@ -62,7 +62,7 @@ class MutableProviderTest extends \Tester\TestCase
 
 	public function testMicroseconds(): void
 	{
-		$tp = new MutableProvider(new DateTimeImmutable('2013-09-14 03:53:21.123456'));
+		$tp = new MutableDateTimeProvider(new DateTimeImmutable('2013-09-14 03:53:21.123456'));
 		$datetime = $tp->getDateTime();
 		$date = $tp->getDate();
 		$time = $tp->getTime();
@@ -78,4 +78,4 @@ class MutableProviderTest extends \Tester\TestCase
 
 }
 
-(new MutableProviderTest())->run();
+(new MutableDateTimeProviderTest())->run();
